issue-35 Added unit tests methods in user service.

+++ src/main/java/tahafurkan/sandbox/usermanagement/service/UserService.java
@@ -48,16 +48,23 @@ public class UserService {
     }
 
     public UserDto update(int id, UserDto user) {
-        UserDto existingUser = get(id);
+        User existingUser = userRepository.findById(id).get();
         existingUser.setName(user.getName());
         existingUser.setLastName(user.getLastName());
         existingUser.setEmail(user.getEmail());
         existingUser.setAddress(user.getAddress());
-        return modelMapper.map(existingUser, UserDto.class);
+        User updateUser = userRepository.save(existingUser);
+        return modelMapper.map(updateUser, UserDto.class);
     }
 
     public void delete(int id) {
-        userRepository.deleteById(id);
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            userRepository.deleteById(id);
+        } else {
+            String msg = String.format(" ID = %d , this ID does not exist  ", id);
+            throw new NoSuchUserExistsException(msg);
+        }
     }
 
     public void checkUsername(String username) {
